Move can_ack_page_complete to canReplyMsg
